refactor naming and structure
