+++ app/javascript/components/Workbench/DatabaseSchemaTree/ColumnItem.tsx
