+++ app/javascript/components/Workbench/DatabaseSchemaTree/DatabaseItem.tsx
