+++ app/javascript/components/Workbench/DatabaseSchemaTree/PlaceholderItem.tsx
