+++ app/javascript/components/Workbench/DatabaseSchemaTree/SchemaItem.tsx
