+++ app/javascript/components/Workbench/DatabaseSchemaTree/TableItem.tsx
