+++ app/javascript/components/Workbench/SqlResult/EmptyResult.tsx
