+++ app/javascript/components/Workbench/SqlResult/ErrorReport.tsx
