+++ app/javascript/components/Workbench/SqlResult/Graph.tsx
