+++ app/javascript/components/Workbench/SqlResult/Graphs/LineGraph.tsx
