+++ app/javascript/components/Workbench/SqlResult/Graphs/LineGraphConfig.tsx
