+++ app/javascript/components/Workbench/SqlResult/Graphs/WordCloud.tsx
