+++ app/javascript/components/Workbench/SqlResult/Graphs/WordCloudConfig.tsx
