+++ app/javascript/components/Workbench/SqlResult/ResultTable.tsx
