+++ app/javascript/components/Workbench/SqlResult/SkippedResult.tsx
