+++ app/javascript/components/Workbench/SqlResult/SqlResult.tsx
