+++ app/javascript/components/Workbench/SqlResults.tsx
